nir: add type alignment support to lower_io
For cl we can have structs with 8/16/32/64 bit scalar types (as well as,
ofc, arrays/structs/etc), which are padded according to 'C' rules.  So
for lowering struct deref's we need to not just consider a field's size,
but also it's alignment.

v2: rename nir_memory_model to nir_io_type_size_align_cb
    remove align and use ALIGN_POT instead